style: move BaseTimeEntity package
