+++ src/main/java/com/wanted/preonboarding/ticket/domain/base/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.ticket.domain;
+package com.wanted.preonboarding.ticket.domain.base;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
